Specify delay as uint16_t to avoid overflow
On 16-bit platforms like ATmega based Arduino boards a delay
of 5*60*1000 = 300'000 is above the max allowed number of an
integer.

Arduino delay expects unsinged long and we have to explicitly
cast bigger delay values to ensure no overflow occurs.

+++ SCD4x_I2C_FRC_Forced_Recalibration_Example/SCD4x_I2C_FRC_Forced_Recalibration_Example.ino
@@ -60,7 +60,7 @@ void setup() {
 
   // wait for 5 minutes to equilibrate sensor to ambient
   Serial.println("# Waiting 5 minutes for equilibration");
-  delay(5 * 60 * 1000);
+  delay((unsigned long)5 * 60 * 1000);
 
   Serial.println("# CO2 values before calibration");
